Apply Kahan algorithm in PSD of each segment

+++ src/freq_response.h
@@ -7,6 +7,6 @@ int get_freq_response (double complex *poles, int npoles,
                        double complex *zeros, int nzeros,
                        double constant, double sampling_rate, int data_samples,
                        double **freq, double complex **freq_response, int flag);
-int remove_response (double complex *data, double complex *freq_response, int data_samples);
+int remove_response (float complex *data, double complex *freq_response, int data_samples);
 
 #endif
